requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
boto3==1.37.30
botocore==1.37.30
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
cssselect2==0.7.0
Django==5.1.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-storages==1.14.6
djangorestframework==3.15.2
html5lib==1.1
idna==3.10
jmespath==1.0.1
lxml==5.3.1
oscrypto==1.3.0
pillow==11.1.0
pillow_heif==0.22.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.3.0
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17
